Add LICENSE
Introducing the LICENSE file to ensure legal protection for contributors and define the permissible uses of this project. The license outlines the terms and conditions under which the project can be used, modified, and distributed. Adhering to open-source principles, this license promotes collaboration while safeguarding the rights of the creator and contributors.